% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePalette.R
\name{MakePalette}
\alias{MakePalette}
\title{Make your palette of colors}
\usage{
MakePalette(photo, n = 4)
}
\arguments{
\item{photo}{Image location path. It can also be a URL address.}

\item{n}{Number of elements to be generated in the color palette. The default value is 4.}
}
\value{
A palette of colors
}
\description{
Creates a color palette from an image, using the KMeans algorithm
}
\examples{
MakePalette(system.file("extdata", "picture02.jpg", package="MakeYourPalette"))
MakePalette(system.file("extdata", "picture04.png", package="MakeYourPalette"))
MakePalette(system.file("extdata", "picture06.png", package="MakeYourPalette"))

}
